requirements.txt
attackcti
azure-identity
azure-mgmt-securityinsight == 1.0.0b2
mitreattack-python
pyyaml
requests
stix2
taxii2-client